requirements.txt
python-dotenv==0.19.2
webdriver-manager==3.5.3
selenium==4.1.3
beautifulsoup4==4.10.0
requests==2.27.1
pandas==1.4.1